// src/openapi/models/document-space-privilege-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * TRON Common API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface DocumentSpacePrivilegeDto
 */
export interface DocumentSpacePrivilegeDto {
    /**
     * 
     * @type {string}
     * @memberof DocumentSpacePrivilegeDto
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof DocumentSpacePrivilegeDto
     */
    type: DocumentSpacePrivilegeDtoTypeEnum;
}

/**
    * @export
    * @enum {string}
    */
export enum DocumentSpacePrivilegeDtoTypeEnum {
    Read = 'READ',
    Write = 'WRITE',
    Membership = 'MEMBERSHIP'
}